+++ Hitai/Dialogs/GenerateKeyPairDialog.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Net.Mail;
+using System.Threading;
 using System.Threading.Tasks;
 using System.Windows.Forms;
 using Hitai.AsymmetricEncryption;
@@ -8,6 +9,9 @@ namespace Hitai.Dialogs
 {
     public partial class GenerateKeyPairDialog : Form
     {
+        private readonly CancellationTokenSource _cancellationTokenSource =
+            new CancellationTokenSource();
+
         private IAsymmetricEncryptionProvider _aep;
 
         public GenerateKeyPairDialog() {
@@ -18,6 +22,8 @@ namespace Hitai.Dialogs
         }
 
         public Keypair Keypair { get; set; }
+
+        public bool Cancelled { get; set; }
 
         private async void butGenerate_Click(object sender, EventArgs e) {
             // validate form
@@ -53,13 +59,21 @@ namespace Hitai.Dialogs
                         break;
                     case 1: // Hitai RSA
                         var hiaep = new HitaiAsymmetricEncryptionProvider();
-                        hiaep.GenerateNewKeypair();
+                        try {
+                            hiaep.GenerateNewKeypair(_cancellationTokenSource.Token);
+                        }
+                        catch (OperationCanceledException) {
+                            Cancelled = true;
+                            return;
+                        }
+
                         _aep = hiaep;
                         break;
                     default:
                         throw new NotImplementedException();
                 }
             });
+            if (Cancelled) return;
             progressBar.Style = ProgressBarStyle.Continuous;
             butGenerate.Enabled = true;
             var passwordCreationDialog = new CreatePasswordDialog();
@@ -77,6 +91,8 @@ namespace Hitai.Dialogs
 
         private void butCancel_Click(object sender, EventArgs e) {
             DialogResult = DialogResult.Cancel;
+            _cancellationTokenSource.Cancel();
+            Cancelled = true;
             Close();
         }