+++ changelog.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
+### Fixed
+- cancellation of Hitai keypair generation now aborts background threads
 
 ## 0.1.0 - 2019-02-02
 - First release